- styled setting and it's responsive - Minified the components and made settings more simple making more sense - Deleted routes and components that were useless - Added routes found to App.js
