+++ client/src/components/Account/Account.js
@@ -1,45 +1,53 @@
 import React, { Component } from 'react';
-import { Link, Route } from 'react-router-dom';
+import { Link } from 'react-router-dom';
+import StripeCheckout from 'react-stripe-checkout';
+import { gql } from 'apollo-boost';
+
 import * as ROUTES from '../../constants/routes';
 import { withAuthorization } from '../Session/session';
-import ProjectList from './Lists/ProjectList';
-import ReviewList from './Lists/ReviewList';
-import Stripe from '../Stripe/Stripe';
 import PasswordChange from '../PasswordChange/PasswordChange';
+import './Account.scss';
+import '../../styles/_globals.scss';
+
+const createSubscriptionMutation = gql`
+  mutation createSubscription($source: String!, $email: String!) {
+    createSubscription(source: $source, email: $email) {
+      id
+      email
+    }
+  }
+`;
+
 
 class Account extends Component {
+  constructor(props) {
+    super(props)
+  }
   render() {
+    const user = this.props.firebase.auth.currentUser;
     return (
-      <div>
-        <div className="accountNav">
-          <ul>
-            <li>
-              <Link to={ROUTES.MY_PROJECTS}>My Projects</Link>
-            </li>
-            <li>
-              <Link to={ROUTES.MY_REVIEWS}>My Reviews</Link>
-            </li>
-            <li>
-              <Link to={ROUTES.STRIPE}>Billing</Link>
-            </li>
-          </ul>
-        </div>
-        <div>
+      <>
+        <div className='settings-container'>
+          <h1>Settings</h1>
           <PasswordChange />
+          <div className='stripe-container'>
+            <h2>Want to Upgrade?</h2>
+            <StripeCheckout
+              className='btn'
+              token={async (token) => {
+                const response = await this.props.mutate({
+                  variables: { source: token.id, email: user.email }
+                });
+              }}
+              stripeKey="pk_test_c80Nc7ujL3MIYgeZj479Sn0H"
+            />
+          </div>
         </div>
-
-        <Route path={ROUTES.MY_PROJECTS} component={ProjectList} />
-        <Route path={ROUTES.MY_REVIEWS} component={ReviewList} />
-        <Route
-          path={ROUTES.STRIPE}
-          render={(props) => (
-            <Stripe {...props} firebase={this.props.firebase} />
-          )}
-        />
-      </div>
+      </>
     );
   }
 }
+
 const condition = (authUser) => authUser;
 
 export default withAuthorization(condition)(Account);
