+++ client/src/styles/card.scss
@@ -14,7 +14,7 @@
   width: 23%;
   margin: 20px 0;
   height: 500px;
-  border: 2px solid #ecf0f1;
+  border: 2px solid #388087;
   text-align: center;
 
   @media (max-width: 600px) {
